# backend/swagger.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: tSNS API
  description: "**tiny-SNS API** is a node-JS express backend providing basic instagram alike functionalities such as registering, following, posting, commenting and expressing how one feels about posts and comments"
  contact:
    name: Jiwoo Jeon
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
paths:
...
  /api/comments:
    post:
      tags:
        - comments
      security:
        - bearerAuth: []
      summary: create a post
      description: append image files (png) to a request
      requestBody:
        content:
          application/json:
            schema:
              properties:
                postID:
                  type: string
                content:
                  type: string

            example:
              postID: p5faf8980f1199903d7e1eff1
              content: user1-main thread content 1

      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/commentObj"

              example:
                reactions: { love: 0, haha: 0, sad: 0, angry: 0 }
                parentComID: null
                numChild: 0
                _id: "c5fae7ca6e76c622ea41cde6f"
                postID: "p5fae7ca5e76c622ea41cde66"
                user: { _id: "5fae7ca4e76c622ea41cde62", nickname: "user1" }
                content: "user1-main thread content 1"
                createdAt: "2020-11-13T12:31:34.239Z"
                updatedAt: "2020-11-13T12:31:34.239Z"
                __v: 0

        "400":
          description: Bad Request, reason in body
        # "500":
        #   description: Internal Error
  /api/comments/{postID}:
    get:
      tags:
        - comments
      security:
        - bearerAuth: []
      summary: get posts on user's home
      description: get "public"/ "follower" level posts of following people and all level ("public"/ "follower"/ "private") posts of the user
      parameters:
        - in: path
          name: postID
          schema:
            type: string
          required: true
          description: ID of the post to remove user's reaction
          example: "p5fa53f99ab999d06f32575e9"
        - in: query
          name: last-created-at
          schema:
            type: string
          description: createdAt time of the last content for paging purposes
        - in: query
          name: num
          schema:
            type: number
          description: number of contents to get
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/postObj"
              example:
                [
                  {
                    _id: "p5fab9973bc99520da2f88967",
                    reactions: { love: 0, haha: 0, sad: 0, angry: 0 },
                    description: "user2_FollowersPost",
                    media: [],
                    level: "followers",
                    user:
                      { _id: "5fab9972bc99520da2f8895e", nickname: "user2" },
                    createdAt: "2020-11-11T07:57:39.013Z",
                    updatedAt: "2020-11-11T07:57:39.013Z",
                    __v: 0,
                    userReaction: null,
                  },
                  {
                    _id: "p5fab9972bc99520da2f88966",
                    reactions: { love: 0, haha: 0, sad: 0, angry: 0 },
                    description: "user2_PublicPost",
                    media: [],
                    level: "public",
                    user:
                      { _id: "5fab9972bc99520da2f8895e", nickname: "user2" },
                    createdAt: "2020-11-11T07:57:38.984Z",
                    updatedAt: "2020-11-11T07:57:38.984Z",
                    __v: 0,
                    userReaction: null,
                  },
                  {
                    _id: "p5fab9972bc99520da2f88965",
                    reactions: { love: 0, haha: 0, sad: 0, angry: 0 },
                    description: "user1_PrivatePost",
                    media: [],
                    level: "private",
                    user:
                      { _id: "5fab9971bc99520da2f8895d", nickname: "user1" },
                    createdAt: "2020-11-11T07:57:38.954Z",
                    updatedAt: "2020-11-11T07:57:38.954Z",
                    __v: 0,
                    userReaction: null,
                  },
                ]

        "400":
          description: Bad Request, reason in body
  # /api/comments/mine:
  #   get:
  #     tags:
  #       - comments
  #     security:
  #       - bearerAuth: []
  #     summary: remove post
  #     parameters:
  #       - in: query
  #         name: last-created-at
  #         schema:
  #           type: string
  #         description: createdAt time of the last content for paging purposes
  #       - in: query
  #         name: num
  #         schema:
  #           type: number
  #         description: number of contents to get
  #     responses:
  #       "200":
  #         description: OK
  #         content:
  #           application/json:
  #             schema:
  #               type: array
  #               items:
  #                 $ref: "#/components/schemas/postObj"
  #             example:
  #               [
  #                 {
  #                   _id: "p5fabef54a7325615d0d5c373",
  #                   reactions: { love: 0, haha: 0, sad: 0, angry: 0 },
  #                   description: "user1_PrivatePost",
  #                   media: [],
  #                   level: "private",
  #                   user:
  #                     { _id: "5fabef53a7325615d0d5c36b", nickname: "user1" },
  #                   createdAt: "2020-11-11T14:04:04.207Z",
  #                   updatedAt: "2020-11-11T14:04:04.207Z",
  #                   __v: 0,
  #                   userReaction: null,
  #                 },
  #                 {
  #                   _id: "p5fabef54a7325615d0d5c372",
  #                   reactions: { love: 0, haha: 0, sad: 0, angry: 0 },
  #                   description: "user1_FollowersPost",
  #                   media: [],
  #                   level: "followers",
  #                   user:
  #                     { _id: "5fabef53a7325615d0d5c36b", nickname: "user1" },
  #                   createdAt: "2020-11-11T14:04:04.174Z",
  #                   updatedAt: "2020-11-11T14:04:04.174Z",
  #                   __v: 0,
  #                   userReaction: null,
  #                 },
  #                 {
  #                   _id: "p5fabef54a7325615d0d5c371",
  #                   reactions: { love: 0, haha: 0, sad: 0, angry: 0 },
  #                   description: "user1_PublicPost",
  #                   media: [],
  #                   level: "public",
  #                   user:
  #                     { _id: "5fabef53a7325615d0d5c36b", nickname: "user1" },
  #                   createdAt: "2020-11-11T14:04:04.141Z",
  #                   updatedAt: "2020-11-11T14:04:04.141Z",
  #                   __v: 0,
  #                   userReaction: null,
  #                 },
  #               ]

  #       "400":
  #         description: Bad Request, reason in body
...
  /api/comments/{(sub)commentID}:
    delete:
      tags:
        - comments
      security:
        - bearerAuth: []
      summary: remove post
      parameters:
        - in: path
          name: (sub)commentID
          schema:
            type: string
          required: true
          description: ID of comment to delete
      responses:
        "204":
          description: No Content

        "400":
          description: Bad Request, reason in body
  /api/comments/react:
    post:
      tags:
        - comments
        - react
      security:
        - bearerAuth: []
      summary: "react to a post"
      description: "show one of love, haha, sad, angry emotions"
      requestBody:
        content:
          application/json:
            schema:
              properties:
                _id:
                  $ref: "#/components/schemas/_id"
                reaction:
                  $ref: "#/components/schemas/reaction"
              example:
                _id: p5fabef54a7325615d0d5c373
                reaction: haha
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/postObj"

              example:
                reactions: { love: 0, haha: 1, sad: 0, angry: 0 }
                userReaction: "haha"

        "400":
          description: Bad Request, reason in body
        "500":
          description: Internal Error
  /api/comments/react/{(sub)commentID}:
    delete:
      tags:
        - comments
        - react
      security:
        - bearerAuth: []
      summary: undo reaction
      description: delete user's reaction
      parameters:
        - in: path
          name: (sub)commentID
          schema:
            type: string
          required: true
          description: ID of the post to remove user's reaction
          example: "p5fa53f99ab999d06f32575e9"
      responses:
        "200":
          description: No Content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/postObj"

              example:
                reactions: { love: 0, haha: 0, sad: 0, angry: 0 }
                userReaction: null

        "400":
          description: Bad Request, reason in body
        # "500":
        #   description: Internal Error

##################################################################
##################################################################
##################################################################
##################################################################
##################################################################
##################################################################
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    nickname:
      type: string
      minimum: 3
...
    _id:
      type: string
      uniqueItems: true
      description: generated by the API
    reaction:
      type: string
      enum:
        - love
        - haha
        - sad
        - angry
    level:
      type: string
      enum:
        - public
        - followers
        - private
      description: view level for posts
...
    userObj:
      properties:
        _id:
          $ref: "#/components/schemas/_id"
        nickname:
          $ref: "#/components/schemas/nickname"
...
    reactionsObj:
      properties:
        love:
          type: integer
        haha:
          type: integer
        sad:
          type: integer
        angry:
          type: integer
    postObj:
      properties:
        description:
          type: string
          maximum: 200
        media:
          type: array
          items:
            type: string
        level:
          $ref: "#/components/schemas/level"
        _id:
          $ref: "#/components/schemas/_id"
        user:
          $ref: "#/components/schemas/userObj"
        reactions:
          $ref: "#/components/schemas/reactionsObj"
        userReaction:
          $ref: "#/components/schemas/reaction"
        createdAt:
          type: string
        updatedAt:
          type: string
    commentObj:
      properties:
        reactions:
          $ref: "#/components/schemas/reactionsObj"
        userReaction:
          $ref: "#/components/schemas/reaction"
        parentComID:
          $ref: "#/components/schemas/_id"
        numChild:
          type: number
        _id:
          $ref: "#/components/schemas/_id"
        postID:
          $ref: "#/components/schemas/_id"
        user:
          $ref: "#/components/schemas/userObj"
        content:
          type: string
          minimum: 1
          maximum: 150
        createdAt:
          type: string
        updatedAt:
          type: string
        "__v":
          type: number
servers:
  # Added by API Auto Mocking Plugin
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/jjw2995/tSNS_API/1.0.0